+++ src/component/button/button.stories.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
-import Button from './index'
+import Button from './button'
 import { observable } from 'mobx'
+import ButtonGroup from './button_group'
 
 const store = observable({
   count: 0,
@@ -21,4 +22,17 @@ storiesOf('Button', module)
   ))
   .add('with mobx', () => (
     <Button onClick={() => store.addCount()}>aaa{store.count}</Button>
+  ))
+  .add('Button Group', () => (
+    <ButtonGroup>
+      <Button
+        className='btn btn-default'
+        onClick={() =>
+          new Promise(resolve => setTimeout(() => resolve(), 2000))
+        }
+      >
+        点击显示 loading
+      </Button>
+      <Button className='btn btn-default'>123123</Button>
+    </ButtonGroup>
   ))
