+++ src/component/button/button_group.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react'
+
+class ButtonGroup extends Component {
+  render() {
+    const { children } = this.props
+    return <>{children}</>
+  }
+}
+export default ButtonGroup
